move controller + entity in new folder named src/controller or src/entity
