+++ src/controller/TestController.class.php
@@ -1,5 +1,5 @@
 <?php
-namespace www\controllers;
+namespace src\controller;
 use core\router\Route;
 
 class TestController {
