+++ src/entity/UserEntity.class.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace src\entity;
+
+class User
+{
+	public ?int $id = null;
+	public ?string $name = null;
+	public ?string $password = null;
+
+	public function __construct()
+	{
+	}
+}
